+++ core-bundle/src/Resources/contao/modules/ModuleCloseAccount.php
@@ -139,7 +139,7 @@ class ModuleCloseAccount extends \Module
 				$this->User->logout();
 
 				// Check whether there is a jumpTo page
-				if (($objJumpTo = $this->objModel->getRelated('jumpTo')) !== null)
+				if (($objJumpTo = $this->objModel->getRelated('jumpTo')) instanceof PageModel)
 				{
 					$this->jumpToOrReload($objJumpTo->row());
 				}