+++ core-bundle/src/Resources/contao/modules/ModuleCustomnav.php
@@ -140,7 +140,7 @@ class ModuleCustomnav extends \Module
 						break;
 
 					case 'forward':
-						if (($objNext = $objModel->getRelated('jumpTo')) !== null)
+						if (($objNext = $objModel->getRelated('jumpTo')) instanceof PageModel)
 						{
 							/** @var PageModel $objNext */
 							$href = $objNext->getFrontendUrl();