+++ core-bundle/src/Resources/contao/modules/ModuleLogin.php
@@ -76,7 +76,7 @@ class ModuleLogin extends \Module
 			else
 			{
 				// Redirect to the jumpTo page
-				if ($this->jumpTo && ($objTarget = $this->objModel->getRelated('jumpTo')) !== null)
+				if ($this->jumpTo && ($objTarget = $this->objModel->getRelated('jumpTo')) instanceof PageModel)
 				{
 					/** @var PageModel $objTarget */
 					$strRedirect = $objTarget->getFrontendUrl();