+++ core-bundle/src/Resources/contao/modules/ModuleLogout.php
@@ -65,7 +65,7 @@ class ModuleLogout extends \Module
 		}
 
 		// Redirect to jumpTo page
-		elseif ($this->jumpTo && ($objTarget = $this->objModel->getRelated('jumpTo')) !== null)
+		elseif ($this->jumpTo && ($objTarget = $this->objModel->getRelated('jumpTo')) instanceof PageModel)
 		{
 			/** @var PageModel $objTarget */
 			$strRedirect = $objTarget->getFrontendUrl();