+++ core-bundle/src/Resources/contao/modules/ModuleSearch.php
@@ -91,7 +91,7 @@ class ModuleSearch extends \Module
 		$this->Template->advanced = ($this->searchType == 'advanced');
 
 		// Redirect page
-		if ($this->jumpTo && ($objTarget = $this->objModel->getRelated('jumpTo')) !== null)
+		if ($this->jumpTo && ($objTarget = $this->objModel->getRelated('jumpTo')) instanceof PageModel)
 		{
 			/** @var PageModel $objTarget */
 			$this->Template->action = $objTarget->getFrontendUrl();